+++ src/Features/ERDDAP/Platform/Observations/Menu/index.tsx
@@ -7,11 +7,10 @@ import Link from "next/link"
 import { Dropdown, DropdownMenu, DropdownToggle, NavItem, NavLink } from "reactstrap"
 
 import { paths } from "Shared/constants"
-import { buildSearchParamsQuery, urlPartReplacer } from "Shared/urlParams"
+import { urlPartReplacer } from "Shared/urlParams"
 
 import { DataType } from "../../../types"
 import { UsePlatformRenderProps } from "../../../hooks/BuoyBarnComponents"
-import { aWeekAgoRounded, formatDate, weeksInFuture } from "Shared/time"
 
 const initialState = {
   dropdownOpen: false,
