+++ src/Features/ERDDAP/Platform/Observations/WaterLevel/WLErddapObservationTable.tsx
@@ -4,15 +4,14 @@
 import React, { useEffect, useState } from "react"
 import { ListGroup, ListGroupItem } from "reactstrap"
 
+import { DatumOffsets } from "Features/ERDDAP/types"
+import { DatumSelector } from "Features/ERDDAP/waterLevel/DatumSelector"
+import { platformName } from "Features/ERDDAP/utils/platformName"
 import { UnitSystem } from "Features/Units/types"
 
 import { UsePlatformRenderProps } from "../../../hooks/BuoyBarnComponents"
 import { currentConditionsTimeseries, filterTimeSeries } from "../../../utils/currentConditionsTimeseries"
-
 import { itemStyle, TableItem } from "../Table/item"
-import { DatumOffsets } from "Features/ERDDAP/types"
-import { DatumSelector } from "Features/ERDDAP/waterLevel/DatumSelector"
-import { platformName } from "Features/ERDDAP/utils/platformName"
 
 interface Props extends UsePlatformRenderProps {
   unitSelector?: React.ReactNode
