+++ src/Features/ERDDAP/types.ts
@@ -30,6 +30,7 @@ export interface PlatformProperties {
   mooring_site_desc: string
   nbdc_site_id?: string
   station_name?: string
+  platform_type: "Tide Station" | "Buoy" | "Overland Flood"
   // uscg_light_letter?: string
   // watch_circle_radius?: number
 }
