+++ src/Features/ERDDAP/waterLevel/hooks.ts
@@ -1,14 +1,9 @@
+import { usePlatforms } from "Features/ERDDAP/hooks"
 import { useDefaultStringQueryParam } from "Shared/urlParams"
-import {
-  aDayAgoRounded,
-  aWeekAgoRounded,
-  daysAgoRounded,
-  daysInFuture,
-  fullBeginningDateIso,
-  formatDate,
-  manuallySetFullEODIso,
-} from "Shared/time"
-import { Datums } from "../types"
+import { aWeekAgoRounded, daysAgoRounded, daysInFuture, formatDate } from "Shared/time"
+
+import { Datums, PlatformFeatureCollection } from "../types"
+import { conditions } from "../utils/conditions"
 
 // Access and set the end time value for water level pages with a reasonable default if platform should show future data
 export function useEndTime(graphFuture: boolean = false): { endTime: Date; setEndTime: (newQuery: string) => void } {
@@ -44,4 +39,26 @@ export function useDatum(): { datum: Datums; setDatum: (newQuery: Datums) => voi
   const { value, setValue } = useDefaultStringQueryParam("datum", Datums.MLLW)
 
   return { datum: (value || Datums.MLLW) as Datums, setDatum: setValue }
+}
+
+export const filterForSensors = (platforms: PlatformFeatureCollection) => {
+  return platforms?.features.filter((p) => {
+    return (
+      p.properties.platform_type === "Tide Station" ||
+      p.properties.platform_type === "Overland Flood" ||
+      // p.properties.readings.find((r) => r.data_type.short_name === "SWL") ||
+      p.properties.attribution[0]?.attribution === "Hohonu" ||
+      conditions.waterLevel.find((c) => p.properties.readings.find((r) => r.data_type.standard_name === c))
+    )
+  })
+}
+
+// Return just the platforms with water level sensors
+export const useWaterLevelPlatforms = () => {
+  const { data } = usePlatforms()
+  if (!data) {
+    return undefined
+  }
+
+  return filterForSensors(data)
 }
