+++ src/Shared/time.ts
@@ -63,7 +63,6 @@ export function formatDate(date) {
  * Get the current date in the formate that <input type="date"> accepts
  * @returns Current date in YYYY-MM-DD format
  */
-
 export function getToday() {
   return formatDate(new Date())
 }
@@ -159,6 +158,9 @@ export function aDayAgoRounded(): Date {
   return dayAgo
 }
 
+/**
+ * Get a date that is a 3 days previous
+ */
 export function threeDaysAgoRounded(): Date {
   const threeDaysAgo = new Date(Date.now() - THREE_DAYS)
 
@@ -167,6 +169,9 @@ export function threeDaysAgoRounded(): Date {
   return threeDaysAgo
 }
 
+/**
+ * Get a date that is a number of days previous
+ */
 export function daysAgoRounded(numberOfDays): Date {
   const daysAgo = new Date(Date.now() - DAY * numberOfDays)
