+++ src/components/PlatformInfo/WaterLevelSensorInfo.tsx
@@ -3,13 +3,12 @@ import React from "react"
 
 import { PlatformAlerts } from "Features/ERDDAP/Platform/Alerts"
 import { ErddapPlatformInfoPanel } from "Features/ERDDAP/Platform/Info"
-
+import { WaterLevelSensorSelector } from "Features/ERDDAP/waterLevel/sensorSelector"
+import { WLErddapObservationTable } from "Features/ERDDAP/Platform/Observations/WaterLevel/WLErddapObservationTable"
 import { UnitSelector, useUnitSystem } from "Features/Units"
 import { aDayAgoRounded } from "Shared/time"
 
 import { PlatformMatchParams } from "./types"
-import { WaterLevelSensorSelector } from "Features/ERDDAP/waterLevel/sensorSelector"
-import { WLErddapObservationTable } from "Features/ERDDAP/Platform/Observations/WaterLevel/WLErddapObservationTable"
 
 /**
  * Display our platform info panel for the select platform.
